+++ 0x07-pointers_arrays_strings/main.h
@@ -3,7 +3,7 @@
 #define MAIN_H
 
 void set_string(char **s, char *to);
-void *_memset(void *s, char b, size_t n);
+void *_memset(void *s, char b, unsigned int n);
 void print_diagums(int *a, int size);
 char *_strpbrk(char *s, char *accept);
 char _putchar(char *c);
